Fixup SQLITE query
The previous patch was trying to fix an issue with the SQLITE upgrade.
This has hidden another issue. The location column was never requested
and then the default value was always set. In the UI, the food was
always displayed as it was in the fridge.
Update the query to fix the issue.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>